Improve CompilerMonad and Parser
Instead of having a dedicated type for CompilerMonad, it makes more sense to
implement the CompilerMonadT monad transformer and making CompilerMonad an
instance of CompilerMonadT with the Identity monad. This makes it easy to ensure
the most CompilerEnvironment helper functions work on all monads transformed
using CompilerMonadT.

With the new compiler monad implementation, it is easy to create a state monad
for the parser.