Implement basic block merging functions
A helper function is implemented that merges two basic blocks together. It
assumes that the caller knows what they're doing and does not check whether the
merge is legal. In addition, a helper function that merges to basic block lists
by merging the last block in the first list with the first block in the second
list is also implemented. These are used when generating MIL for Block
statements, where it is always safe to merge the generated basic block lists.