Improve output printing
Instead of running the various compilations, getting the output from them, and
printing it using separate IO monad, wrap some of the compiler monads *inside*
an IO monad. This allows most details about output printing to be hidden away.
It also requires that several services be generalized to operate on a
CompilerMonadT instead of just CompilerMonad.