Generalize encoding of generated target code
Instead of defining a special encoding function for in the code generator,
create a type class that defines the basic interface for encoding generated
code. This is then used to write the encoded output from the compiler directly
to a file.

Also add the RSM csh implementation as part of the file content header.