Implementing simple Happy parser
As part of this change, the Alex lexer is re-implemented to follow the new
grammar. In addition, the AST is also refactored to fit the Happy model better.
Specifically, instead of embedding the position of nodes in the nodes
themselves, a special node is created that simply hold other AST nodes plus an
AlexPosn to store the position. This reduces some of the code duplication.